Add files via upload
This project aims to create a school bus monitoring system that improves safety and convenience for students and parents. Using RFID technology and GPS, the system tracks when a student boards the bus, sending real-time notifications and location updates to parents. It utilizes an Arduino Mega and ESP8266 for managing RFID scans and cloud connectivity. This solution addresses parental concerns about their child’s whereabouts and ensures efficient communication and monitoring for school transportation.